requirements.txt
streamlit
pandas
numpy
pillow
joblib
scikit-learn
easyocr
torch
pytesseract
opencv-python-headless
